Title: 합과 평균, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240606/합과 평균/sum-and-avg.java
@@ -9,6 +9,6 @@ public class Main {
         int b = sc.nextInt();
 
         System.out.println(a + b);
-        System.out.printf(".1f", (a + b)/2);        
+        System.out.printf(".1f", double(a + b)/2);        
     }
 }
